+++ Project-2.Tests/Property/TestProperty.cs
@@ -1,0 +1,13 @@
+using Project_2.Models;
+using Project_2.Services.Services;
+using Project_2.Data;
+using Project_2.Models.DTOs;
+using Moq;
+
+namespace Project_2.Tests
+{
+    public class TestProperty
+    {
+
+    }
+}
